Update D2_Influence_of_Social_media__Ads_.ipynb
In this project, I have performed EDA (Visual and Statistical) to get various insights from the data set and further developed and analyzed various ML models to predict that whether or not the customer purchased the product based on the Social media Ads. Further fine-tuning of Models is also done to improve the accuracy score.